Separate parsers from control.utilities into control.parsers
